+++ source-code/clj/x/server_router/route_handler/engine.clj
@@ -17,7 +17,7 @@
 (def DEFAULT-CORE-JS "app.js")
 
 ; @constant (keywords in vector)
-(def CLIENT-ROUTE-KEYS [:client-event :core-js :on-leave-event :restricted? :route-parent :route-template])
+(def CLIENT-ROUTE-KEYS [:client-event :core-js :on-leave-event :restricted? :route-template])
 
 ; @constant (keywords in vector)
 (def SERVER-ROUTE-KEYS [:get :core-js :post :restricted? :route-template :server-event])