+++ source-code/clj/x/server_router/route_handler/events.clj
@@ -118,7 +118,6 @@
   ;    Default: route-handler.engine/DEFAULT-CORE-JS
   ;   :restricted? (boolean)(opt)
   ;    Default: false
-  ;   :route-parent (string)(opt)
   ;   :route-template (string)
   ;   :client-event (metamorphic-event)(opt)
   ;    Az útvonal meghívásakor a kliens-oldalon megtörténő esemény