+++ source-code/cljs/app_plugins/item_lister/subs.cljs
@@ -244,16 +244,6 @@
   [db [_ extension-id item-namespace]]
   (r get-meta-item db extension-id item-namespace :sortable?))
 
-(defn set-title?
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  ;
-  ; @param (keyword) extension-id
-  ; @param (keyword) item-namespace
-  ;
-  ; @return (boolean)
-  [db [_ extension-id item-namespace]]
-  (r route-handled? db extension-id item-namespace))
-
 (defn get-description
   ; WARNING! NON-PUBLIC! DO NOT USE!
   ;