+++ source-code/cljs/x/app_router/route_handler/subs.cljs
@@ -252,15 +252,6 @@
   (let [current-route-string (r get-current-route-string db)]
        (uri/uri->fragment current-route-string)))
 
-(defn get-current-route-parent
-  ; @usage
-  ;  (r router/get-current-route-parent db)
-  ;
-  ; @return (string)
-  [db _]
-  (let [current-route-id (r get-current-route-id db)]
-       (get-in db [:router :route-handler/client-routes current-route-id :route-parent])))
-
 
 
 ;; -- Handle subscriptions ----------------------------------------------------
@@ -376,10 +367,6 @@
 ; @usage
 ;  [:router/get-current-route-fragment]
 (a/reg-sub :router/get-current-route-fragment get-current-route-fragment)
-
-; @usage
-;  [:router/get-current-route-parent]
-(a/reg-sub :router/get-current-route-parent get-current-route-parent)
 
 ; @usage
 ;  [:router/at-home?]